Import raw multi-channel accelerometer data stored in mHealth Specification

import_mhealth_csv imports the raw multi-channel accelerometer data stored in mHealth Specification. Note that this function will fail when loading data that have size too big to fit in the memory. For large data file, please use import_mhealth_csv_chunked to load.

import_mhealth_csv(filepath)

Arguments

filepath
string. The filepath of the input data.

Value

dataframe. The imported multi-channel accelerometer signal, with the first column being the timestamps in POSXlct format, and the rest columns being accelerometer values in \(g\) unit.

How is it used in MIMS-unit algorithm?

This function is a File IO function that is used to import data stored in mHealth Specification during algorithm validation.

See also

Examples

  default_ops = options()
  options(digits.secs=3)
  # Use the sample mhealth csv file provided by the package
  filepath = system.file('extdata', 'mhealth.csv', package='MIMSunit')
  filepath

#> [1] "/home/runner/work/MIMSunit/MIMSunit/renv/library/R-3.6/x86_64-pc-linux-gnu/MIMSunit/extdata/mhealth.csv"

  # Load the file
  df = import_mhealth_csv(filepath)

  # Check loaded file
  head(df)

#>         HEADER_TIME_STAMP     X     Y     Z
#> 1 2017-03-16 12:25:49.986 0.417 1.277 0.152
#> 2 2017-03-16 12:25:50.000 0.396 1.249 0.161
#> 3 2017-03-16 12:25:50.013 0.373 1.224 0.179
#> 4 2017-03-16 12:25:50.025 0.355 1.182 0.179
#> 5 2017-03-16 12:25:50.036 0.341 1.148 0.179
#> 6 2017-03-16 12:25:50.049 0.326 1.117 0.161

  # Check more
  summary(df)

#>  HEADER_TIME_STAMP                      X                Y        
#>  Min.   :2017-03-16 12:25:49.99   Min.   :-0.387   Min.   :0.378  
#>  1st Qu.:2017-03-16 12:26:05.60   1st Qu.: 0.220   1st Qu.:0.839  
#>  Median :2017-03-16 12:26:21.23   Median : 0.385   Median :0.962  
#>  Mean   :2017-03-16 12:26:21.23   Mean   : 0.392   Mean   :1.042  
#>  3rd Qu.:2017-03-16 12:26:36.84   3rd Qu.: 0.548   3rd Qu.:1.261  
#>  Max.   :2017-03-16 12:26:52.48   Max.   : 1.246   Max.   :1.678  
#>        Z           
#>  Min.   :-1.01400  
#>  1st Qu.:-0.05200  
#>  Median : 0.09700  
#>  Mean   : 0.02415  
#>  3rd Qu.: 0.16100  
#>  Max.   : 0.40200  

  # Restore default options
  options(default_ops)