The mean and standard deviation of accelerometer summary measure for different acceleration data summary algorithms and for different devices.

A dataframe contains the mean and standard deviation of accelerometer summary measured at different frequencies for the raw accelerometer signals from different devices collected from on a standard elliptical shaker.

measurements_different_devices

Format

A data frame with 235 rows and 8 variables:

DEVICE

The name of different devices, in character

GRANGE

The dynamic range of the device in g, in number

SR

The sampling rate in Hz of the device, in number

TYPE

Accelerometer summary algorithm name, in character

HZ

The frequency of the elliptical shaker, in number

NAME

An alternative name that is friendly for plotting for devices, in character

mean

The mean values of accelerometer summary measure, in number

sd

The standard deviation values of accelerometer summary measure, in number

Source

https://github.com/qutang/MIMSunit-dataset-shaker