Run shiny app to compute MIMSunit values from files

shiny_app runs a local shiny app that provides a user friendly interface to compute mims unit values from files stored in mhealth or actigraph format.

shiny_app(options = list())

Arguments

options
The options passed to shinyApp.

How is it used in MIMS-unit algorithm?

This provides a user friendly graphical interface to load local files, call mims_unit_from_files and display the results as an interactive graph.

See also

Examples

  if (FALSE) {
  shiny_app()
  }