The statistics of accelerometer data over epoch windows for SPADES Lab dataset.

A dataset containing the statistics of accelerometer data over epoch windows for SPADES LAB dataset.

spades_lab_stats

Format

A data frame with 176092 rows and 16 variables:

ACTIVITY_NAME

name of different activities, in chr

LABEL_NAME

label name of each activity, in chr

SEGMENT

the index number of epoch window, in int

HEADER_TIME_STAMP

start time of the epoch window, in POSIXct

START_TIME

start time of the epoch window, in POSIXct

STOP_TIME

stop time of the epoch window, in POSIXct

AVERAGE_AMP

the average value of vector magnitude, in num

SD_AMP

the standard deviation of vector magnitude, in num

AVERAGE_DOMINANT_FREQ

the average value of dominant frequency, in num

SD_DOMINANT_FREQ

the standard deviation of dominant frequency, in num

AVERAGE_DOMINANT_FREQ_SECOND

the average value of second dominant frequency, in num

SD_DOMINANT_FREQ_SECOND

the standard deviation of second dominant frequency, in num

AVERAGE_DOMINANT_FREQ_THIRD

the average value of third dominant frequency, in num

SD_DOMINANT_FREQ_THIRD

the standard deviation of third dominant frequency, in num

LOCATION

The on-body placement of the accelerometer, in chr

SUBJECT

The subject IDs, in chr