Aggregate over epoch to get estimated accelerometer orientation.

aggregate_for_orientation returns a dataframe with accelerometer orientations estimated by Mizell, 2003 over each epoch (see compute_orientation). The epoch start time will be used as timestamp in the first column.

aggregate_for_orientation(df, epoch, estimation_window = 2,
  unit = "deg")

Arguments

df
dataframe. Input accelerometer data in mhealth format. First column should be timestamps in POSIXt format.

epoch
string. Any format that is acceptable by argument breaks in method cut.POSIXt.For example, "1 sec", "1 min", "5 secs", "10 mins".

estimation_window
number. Duration in seconds to be used to estimate orientation within each epoch. Default is 2 (seconds), as suggested by Mizell, 2003.

unit
string. The unit of orientation angles. Can be "deg" (degree) or "rad" (radian). Default is "deg".

Value

dataframe. The returned dataframe will have the same format as input dataframe.

Details

This function accepts a dataframe (in mhealth accelerometer data format) and computes the estimated acclerometer orientations (in x, y, and z angles) over each fixed epoch. The returned dataframe will have the same format as input dataframe, including four columns, and have the same datetime format as input dataframe in the timestamp column. The orientation estimation method used in the function is based on Mizell, 2003.

Note

If epoch argument is not provided or is NULL, the function will treat the input dataframe as a single epoch.

How is it used in mims-unit algorithm?

This function is used in mims-unit algorithm after extrapolation (extrapolate). The extrapolated signal will be estimated to get orientation angles using this function.

See also