Export accelerometer data in Actilife RAW CSV format

export_to_actilife exports the input dataframe as a csv file that is compatible with Actilife.

export_to_actilife(df, filepath, actilife_version = "6.13.3",
  firmware_version = "1.6.0")

Arguments

df
dataframe. Input accelerometer data. The first column is timestamp in POSXlct format, and the rest columns are accelerometer values in g (\(9.81m/s^2\)).

filepath
string. The output filepath.

actilife_version
string. The Actilife version number to be added to the header. Default is "6.13.3", that was used by the algorithm during development.

firmware_version
string. The firmware version number to be added to the header. This is supposed to be the firmware version of the Actigraph devices. We did not see any usage of the number during the computation of Actigraph counts by Actilife, so it may be set with an arbitrary version code seen in any Actigraph devices. We use default version code "1.6.0".

Value

No return value.

Details

This function takes an input accelerometer dataframe and exports it in Actilife RAW CSV format with a prepended a madeup header. The exported file csv file has compatible header, column names, timestamp format with Actilife and can be imported directly into Actilife software.

How is it used in MIMS-unit algorithm?

This function is an utility function that was used to convert validation data into Actilife RAW CSV format so that we can use Actilife to compute Actigraph counts values for these data.