Plot MIMS unit values or raw signal using dygraphs interactive plotting library.

generate_interactive_plot plots MIMS unit values or raw signal using dygraphs interactive plotting library.

generate_interactive_plot(df, y_label, value_cols = c(2, 3, 4))

Arguments

df
data.frame.The dataframe storing MIMS unit values or raw accelerometer signal. The first column should be timestamps.

y_label
str. The label name to be put on the y axis.

value_cols
numerical vector. The indices of columns storing values, typically starting from the second column. The default is `c(2,3,4)`.

Value

A dygraphs graph object. When showing, the graph will be plotted in a html widgets in an opened browser.

See also