Compute activity counts using [citation] algorithm

Compute activity counts using [citation] algorithm

mims_unit(df, breaks = "5 sec", range, noise_level = 0.03, k = 0.05,
  spar = 0.6, filter_type = "butter", cutoffs = c(0.2, 5),
  integration = "trapz", axes = c(2, 3, 4), use_extrapolation = TRUE,
  use_interpolation = TRUE, use_filtering = TRUE,
  combination = "sum", vm_after_extrapolation = FALSE,
  allow_truncation = TRUE, output_per_axis = FALSE,
  output_orientation = FALSE, breaks_for_orientation = NULL,
  before_df = NULL, after_df = NULL)

Arguments

df
input sensor data frame that fits the mhealth specification

breaks
epoch to compute counts on

range
dynamic range of the device (two element vector)

noise_level
noise level should be between 0.01 and 0.1

k
neighborhood duration for extrapolation in seconds, default is 0.05 seconds

spar
smoothing parameter for extrapolation, default is 0.6

cutoffs
cut off frequencies to be used in filtering, default is 0.2Hz and 5Hz. If choosing bessel, the low pass cut off would be multipled by 2 when being used.

integration
the integration method to be used: 'trapz', 'absoluteMeanByPoints'.

allow_truncation
use zero truncation or not, default is TRUE

resample
set 0 to not use resampling, otherwise set to the desired sampling rate in numerical, e.g 40 for 40Hz. Default is 50.

filter
'butter', 'ellip', 'bessel'

Value

Will return NA if a segment break is too short to have enough samples to compute AUC