Estimates sensor orientation

sensor_orientations estimates the orientation angles for the input multi-channel accelerometer signal. The input signal can be from devices of any sampling rate and dynamic range. Please refer to function compute_orientation for the implementation of the estimation algorithm.

sensor_orientations(
  df,
  before_df = NULL,
  after_df = NULL,
  epoch = "5 sec",
  dynamic_range,
  st = NULL
)

Arguments

df
dataframe. Input multi-channel accelerometer signal.

before_df
dataframe. The multi-channel accelerometer signal comes before the input signal to be prepended to the input signal during computation. This is used to eliminate the edge effect during extrapolation and filtering. If it is NULL, algorithm will run directly on the input signal. Default is NULL.

after_df
dataframe. The multi-channel accelerometer signal comes after the input signal to be append to the input signal. This is used to eliminate the edge effect during extrapolation and filtering. If it is NULL, algorithm will run directly on the input signal. Default is NULL.

epoch
string. Any format that is acceptable by argument breaks in method cut.POSIXt.For example, "1 sec", "1 min", "5 sec", "10 min". Default is "5 sec".

dynamic_range
numerical vector. The dynamic ranges of the input signal. Should be a 2-element numerical vector. c(low, high), where low is the negative max value the device can reach and high is the positive max value the device can reach.

st
character or POSIXct timestamp. An optional start time you can set to force the epochs generated by referencing this start time. If it is NULL, the function will use the first timestamp in the timestamp column as start time to generate epochs. This is useful when you are processing a stream of data and want to use a common start time for segmenting data. Default is NULL.

Value

dataframe. The orientation dataframe. The first column is the start time of each epoch in POSIXct format. The second to fourth columns are the orientation angles.

Note

This function interpolates and extrapolates the signal before estimating the orientation angles.

before_df and after_df are often set when the accelerometer data are divided into files of smaller chunk.

How is it used in MIMS-unit algorithm?

This is not included in the official MIMS-unit algorithm nor the manuscript, but we found it is useful to know the sensor orientations in addition to the summary of movement.

See also

Examples

  # Use sample data for testing
  df = sample_raw_accel_data

  # compute sensor orientation angles
  sensor_orientations(df, epoch = '5 min', dynamic_range=c(-8, 8))

#> ================================================

#> Warning: NaNs produced

#> ================================

#>         HEADER_TIME_STAMP    X_ANGLE  Y_ANGLE   Z_ANGLE
#> 1 2016-01-15 11:00:00.013  13.900844 86.12948 100.85193
#> 2 2016-01-15 11:05:00.003   9.662081 93.65115  96.41904
#> 3 2016-01-15 11:10:00.003 102.409417 92.06482  24.60815
#> 4 2016-01-15 11:15:00.003        NaN      NaN       NaN

  # compute sensor orientation angles with different epoch length
  output = sensor_orientations(df, epoch = '15 sec', dynamic_range=c(-8, 8))

#> ================================================

#> Warning: NaNs produced

#> Warning: NaNs produced

#> ================================

  head(output)

#>         HEADER_TIME_STAMP  X_ANGLE  Y_ANGLE   Z_ANGLE
#> 1 2016-01-15 11:00:00.013 10.91751 91.06634 100.74475
#> 2 2016-01-15 11:00:15.003 10.49971 91.29207 100.10203
#> 3 2016-01-15 11:00:30.003 10.46272 91.77603 100.01999
#> 4 2016-01-15 11:00:45.003 10.08617 90.74729  99.82672
#> 5 2016-01-15 11:01:00.003 13.69015 85.14183 101.91828
#> 6 2016-01-15 11:01:15.003 16.71821 79.25918 102.21465